Cleanup and Restructuring to Enable CocoaPods Integration
The structure of the framework was not conducive to allowing use
in CocoaPods. This commit restructures the file layout, adds appropriate
imports to necessary files, and creates a modular Podspec to allow
for integration of only individual components as needed by developers.

In addition, it removes unnecessary customizations of SVProgressHUD,
and includes SVProgressHUD as a dependency to avoid namespace collision
in projects that already include SVProgressHUD.

This places the onus on the developer to either utilize CocoaPods,
or manually include SVProgressHUD to satisfy all dependencies.
As such, this has been noted in the README.